agentic_ritual_engine/requirements.txt
fastapi==0.110.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.6.3
requests==2.31.0
pdf2image==1.17.0
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
beautifulsoup4==4.12.3
PyYAML==6.0.1
Unidecode==1.3.8
python-slugify==8.0.4
pytesseract==0.3.10
PyMuPDF==1.26.4
tqdm==4.66.2
jinja2==3.1.3
typer==0.9.0
uvicorn[standard]==0.27.1
astral==3.2
networkx==3.2.1
matplotlib==3.8.2
streamlit==1.31.1
